+++ osf/management/commands/populate_impact_institution_metrics.py
@@ -4,40 +4,15 @@ from django.core.management.base import BaseCommand
 
 from osf.metrics import (
     InstitutionProjectCounts,
-    PreprintView,
-    PreprintDownload,
     UserInstitutionProjectCounts,
 )
 
-from osf.models import Institution, OSFUser, Preprint
+from osf.models import Institution, OSFUser
 
 
 # Edit the following guids based off your local environment
-USERS = [OSFUser.load('abd53u'), OSFUser.load('hy84n'), OSFUser.load('7zyg2')]
-PREPRINTS = [Preprint.load('far32'), Preprint.load('67bzg')]
+USERS = [OSFUser.load('bd53u'), OSFUser.load('hy84n'), OSFUser.load('7zyg2')]
 INSTITUTION = Institution.load('cos')
-
-def populate_preprint_metrics(dates):
-    for date in dates:
-        for user in USERS:
-            for preprint in PREPRINTS:
-                preprint_view_count = int(10 * random())
-                PreprintView.record_for_preprint(
-                    preprint=preprint,
-                    user=user,
-                    path=preprint.primary_file.path,
-                    timestamp=date,
-                    count=preprint_view_count
-                )
-
-                preprint_download_count = int(10 * random())
-                PreprintDownload.record_for_preprint(
-                    preprint=preprint,
-                    user=user,
-                    path=preprint.primary_file.path,
-                    timestamp=date,
-                    count=preprint_download_count
-                )
 
 
 def populate_institution_metrics(dates):
@@ -71,22 +46,9 @@ class Command(BaseCommand):
 
     def add_arguments(self, parser):
         super(Command, self).add_arguments(parser)
-        parser.add_argument(
-            '--institutions',
-            action='store_true',
-            help='Only populate institutions related date'
-        )
-        parser.add_argument(
-            '--preprints',
-            action='store_true',
-            help='Only populate preprints related date'
-        )
 
     def handle(self, *args, **options):
         today = dt.datetime.today()
         last_seven_days = [(today - dt.timedelta(days=num_days)) for num_days in range(0, 7)]
 
-        if not options['preprints']:
-            populate_institution_metrics(last_seven_days)
-        if not options['institutions']:
-            populate_preprint_metrics(last_seven_days)
+        populate_institution_metrics(last_seven_days)
